renamed input to adapter
