+++ stwo_cairo_prover/crates/adapted_prover/src/main.rs
@@ -2,11 +2,11 @@ use std::path::PathBuf;
 use std::process::ExitCode;
 
 use clap::Parser;
+use stwo_cairo_prover::adapter::vm_import::{adapt_vm_output, VmImportError};
+use stwo_cairo_prover::adapter::ProverInput;
 use stwo_cairo_prover::cairo_air::{
     prove_cairo, verify_cairo, CairoVerificationError, ConfigBuilder,
 };
-use stwo_cairo_prover::input::vm_import::{adapt_vm_output, VmImportError};
-use stwo_cairo_prover::input::ProverInput;
 use stwo_cairo_utils::binary_utils::run_binary;
 use stwo_prover::core::prover::ProvingError;
 use stwo_prover::core::vcs::blake2_merkle::Blake2sMerkleChannel;
