+++ stwo_cairo_prover/crates/prover/src/adapter/memory.rs
