+++ stwo_cairo_prover/crates/prover/src/adapter/mod.rs
