+++ stwo_cairo_prover/crates/prover/src/cairo_air/air.rs
@@ -31,10 +31,10 @@ use super::range_checks_air::{
     RangeChecksInteractionClaim, RangeChecksInteractionClaimGenerator,
     RangeChecksInteractionElements,
 };
+use crate::adapter::ProverInput;
 use crate::components::memory::{memory_address_to_id, memory_id_to_big};
 use crate::components::{verify_bitwise_xor_9, verify_instruction};
 use crate::felt::split_f252;
-use crate::input::ProverInput;
 use crate::relations;
 
 #[derive(Serialize, Deserialize)]
