+++ stwo_cairo_prover/crates/prover/src/components/memory/memory_address_to_id/prover.rs
@@ -16,12 +16,12 @@ use stwo_prover::core::poly::circle::{CanonicCoset, CircleEvaluation};
 use stwo_prover::core::poly::BitReversedOrder;
 
 use super::component::{Claim, InteractionClaim, MEMORY_ADDRESS_TO_ID_SPLIT};
+use crate::adapter::memory::Memory;
 use crate::cairo_air::preprocessed::Seq;
 use crate::components::memory_address_to_id::component::{
     N_ID_AND_MULT_COLUMNS_PER_CHUNK, N_TRACE_COLUMNS,
 };
 use crate::components::utils::AtomicMultiplicityColumn;
-use crate::input::memory::Memory;
 use crate::relations;
 
 pub type PackedInputType = PackedM31;
@@ -219,9 +219,9 @@ mod tests {
     use itertools::Itertools;
     use stwo_prover::core::fields::m31::{BaseField, M31};
 
+    use crate::adapter::memory::{MemoryBuilder, MemoryConfig};
+    use crate::adapter::vm_import::MemoryEntry;
     use crate::components::memory::memory_address_to_id;
-    use crate::input::memory::{MemoryBuilder, MemoryConfig};
-    use crate::input::vm_import::MemoryEntry;
 
     #[test]
     fn test_memory_multiplicities() {
