+++ stwo_cairo_prover/crates/prover/src/components/memory/memory_id_to_big/component.rs
@@ -13,8 +13,8 @@ use stwo_prover::core::fields::secure_column::SECURE_EXTENSION_DEGREE;
 use stwo_prover::core::pcs::TreeVec;
 use stwo_prover::relation;
 
+use crate::adapter::memory::LARGE_MEMORY_VALUE_ID_BASE;
 use crate::cairo_air::preprocessed::{PreProcessedColumn, Seq};
-use crate::input::memory::LARGE_MEMORY_VALUE_ID_BASE;
 use crate::relations;
 
 // TODO(AlonH): Make memory size configurable.
