+++ stwo_cairo_prover/crates/prover/src/components/verify_instruction/prover.rs
@@ -33,11 +33,11 @@ use stwo_prover::core::utils::bit_reverse_coset_to_circle_domain_order;
 use stwo_prover::core::vcs::blake2_merkle::{Blake2sMerkleChannel, Blake2sMerkleHasher};
 
 use super::component::{Claim, InteractionClaim};
+use crate::adapter::decode::{deconstruct_instruction, Instruction};
 use crate::components::utils::pack_values;
 use crate::components::{
     memory_address_to_id, memory_id_to_big, range_check_4_3, range_check_7_2_5,
 };
-use crate::input::decode::{deconstruct_instruction, Instruction};
 use crate::relations;
 
 pub type InputType = (M31, [M31; 3], [M31; 15]);
