+++ stwo_cairo_prover/crates/prover/src/lib.rs
@@ -8,10 +8,10 @@
 // TODO(Ohad): remove.
 #![allow(clippy::too_many_arguments)]
 
+pub mod adapter;
 pub mod cairo_air;
 pub mod components;
 mod felt;
-pub mod input;
 pub mod relations;
 
 pub use stwo_prover;
