+++ stwo_cairo_prover/crates/run_and_prove/src/main.rs
@@ -2,11 +2,11 @@ use std::path::PathBuf;
 use std::process::ExitCode;
 
 use clap::Parser;
+use stwo_cairo_prover::adapter::plain::adapt_finished_runner;
+use stwo_cairo_prover::adapter::vm_import::VmImportError;
 use stwo_cairo_prover::cairo_air::{
     prove_cairo, verify_cairo, CairoVerificationError, ConfigBuilder,
 };
-use stwo_cairo_prover::input::plain::adapt_finished_runner;
-use stwo_cairo_prover::input::vm_import::VmImportError;
 use stwo_cairo_utils::binary_utils::run_binary;
 use stwo_cairo_utils::vm_utils::{run_vm, VmArgs, VmError};
 use stwo_prover::core::prover::ProvingError;
